bpf: Introduce from_tunnel field to ctmap value
Introduce a new field from_tunnel to ctmap value which indicates the
connection is established over tunnel. That means the reply traffic to
the connection always needs to go overlay even if the destination IP
address is reachable through underlay.

The use case is to avoid assymetric traffic when the packet is
inter-cluster SNAT'ed. In the current assumption of overlapping PodCIDR
support for Cluster Mesh, inter-cluster SNAT uses NodeIP as a source
address, thus it's reachable through underlay.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>