ci: Automate generation and update of docs-builder image
The docs-builder image is used for the CI wofkflow for testing
documentation. Each time we change the list of requirements (Sphinx
dependencies etc.) or the Dockerfile for that image, we have to generate
a new image to use in the CI workflow.

So far these images have been generated and pushed to Docker manually
(thanks Joe!). But this is not fun or practical [citation needed]. This
commit introduces a new CI workflow to generate and push a new image
when relevant, and to update the documentation workflow with the new tag
and digest.

For Pull Requests from branches of the cilium/cilium repository, we can
add a new commit directly to update the image. For branches from
external forks, we require an additional step from the PR author to
update the image reference.

Related: #24019
Signed-off-by: Quentin Monnet <quentin@isovalent.com>